fix: split simple_tui and knowledge_builder under 400L limit
