+++ crates/amplihack-utils/src/simple_tui.rs
@@ -362,98 +362,10 @@ impl SimpleTUITester {
 
 // ---------------------------------------------------------------------------
 // Internal command helpers
-// ---------------------------------------------------------------------------
 
-/// Outcome of a single subprocess invocation.
-enum CmdOutcome {
-    Success(String),
-    Failed(String),
-    Timeout,
-    Error(String),
-}
+// Subprocess helpers extracted to simple_tui_runner.rs.
+use crate::simple_tui_runner::{CmdOutcome, command_exists_on_path, run_command_with_timeout};
 
-/// Run a command with a wall-clock timeout.
-fn run_command_with_timeout(args: &[&str], timeout: Duration, cwd: Option<&Path>) -> CmdOutcome {
-    if args.is_empty() {
-        return CmdOutcome::Error("empty argument list".into());
-    }
-
-    let mut cmd = Command::new(args[0]);
-    cmd.args(&args[1..])
-        .stdout(std::process::Stdio::piped())
-        .stderr(std::process::Stdio::piped())
-        .env("CI", "true")
-        .env("DEBIAN_FRONTEND", "noninteractive");
-
-    if let Some(dir) = cwd {
-        cmd.current_dir(dir);
-    }
-
-    let child = match cmd.spawn() {
-        Ok(c) => c,
-        Err(e) => return CmdOutcome::Error(e.to_string()),
-    };
-
-    // Wait with timeout via a polling loop.
-    wait_with_timeout(child, timeout)
-}
-
-/// Poll a child process until it exits or `timeout` elapses.
-fn wait_with_timeout(mut child: std::process::Child, timeout: Duration) -> CmdOutcome {
-    let start = Instant::now();
-    let poll_interval = Duration::from_millis(50);
-
-    loop {
-        match child.try_wait() {
-            Ok(Some(status)) => {
-                let stdout = child
-                    .stdout
-                    .take()
-                    .map(|mut r| {
-                        let mut s = String::new();
-                        std::io::Read::read_to_string(&mut r, &mut s).ok();
-                        s
-                    })
-                    .unwrap_or_default();
-                let stderr = child
-                    .stderr
-                    .take()
-                    .map(|mut r| {
-                        let mut s = String::new();
-                        std::io::Read::read_to_string(&mut r, &mut s).ok();
-                        s
-                    })
-                    .unwrap_or_default();
-
-                return if status.success() {
-                    CmdOutcome::Success(stdout)
-                } else {
-                    CmdOutcome::Failed(stderr)
-                };
-            }
-            Ok(None) => {
-                if start.elapsed() >= timeout {
-                    let _ = child.kill();
-                    let _ = child.wait();
-                    return CmdOutcome::Timeout;
-                }
-                std::thread::sleep(poll_interval);
-            }
-            Err(e) => return CmdOutcome::Error(e.to_string()),
-        }
-    }
-}
-
-/// Return `true` if `name` resolves via `which`.
-fn command_exists_on_path(name: &str) -> bool {
-    Command::new("which")
-        .arg(name)
-        .stdout(std::process::Stdio::null())
-        .stderr(std::process::Stdio::null())
-        .status()
-        .map(|s| s.success())
-        .unwrap_or(false)
-}
 
 // ---------------------------------------------------------------------------
 // Factory / convenience helpers
@@ -482,154 +394,7 @@ pub fn create_amplihack_test(test_id: impl Into<String>, args: &str) -> TUITestC
 // Tests
 // ---------------------------------------------------------------------------
 
+
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_result_passed_constructor() {
-        let r = TestResult::passed("t1", 1.5, "ok");
-        assert!(r.is_passed());
-        assert_eq!(r.status, "passed");
-        assert_eq!(r.test_id, "t1");
-    }
-
-    #[test]
-    fn test_result_failed_constructor() {
-        let r = TestResult::failed("t2", 0.1, "boom");
-        assert!(!r.is_passed());
-        assert_eq!(r.status, "failed");
-        assert_eq!(r.message, "boom");
-    }
-
-    #[test]
-    fn test_result_serde_roundtrip() {
-        let r = TestResult::passed("ser", 2.0, "ok");
-        let json = serde_json::to_string(&r).expect("serialize");
-        let r2: TestResult = serde_json::from_str(&json).expect("deserialize");
-        assert_eq!(r, r2);
-    }
-
-    #[test]
-    fn tui_test_case_default_timeout() {
-        let tc = TUITestCase::new("tc", "my test", vec!["echo hi".into()]);
-        assert_eq!(tc.timeout, 10);
-    }
-
-    #[test]
-    fn tui_test_case_custom_timeout() {
-        let tc = TUITestCase::with_timeout("tc", "my test", vec!["echo hi".into()], 30);
-        assert_eq!(tc.timeout, 30);
-    }
-
-    #[test]
-    fn tui_test_case_serde_roundtrip() {
-        let tc = TUITestCase::new("tc", "test", vec!["ls".into()]);
-        let json = serde_json::to_string(&tc).expect("serialize");
-        let tc2: TUITestCase = serde_json::from_str(&json).expect("deserialize");
-        assert_eq!(tc, tc2);
-    }
-
-    #[test]
-    fn ci_detection_respects_env() {
-        // In test environments CI is typically set, so just verify the function runs.
-        let _ = is_ci_environment();
-    }
-
-    #[test]
-    fn tester_add_and_count() {
-        let dir = tempfile::tempdir().expect("tempdir");
-        let mut tester = SimpleTUITester::new(dir.path()).expect("new");
-        assert_eq!(tester.test_count(), 0);
-
-        tester.add_test(TUITestCase::new("a", "A", vec!["echo a".into()]));
-        tester.add_test(TUITestCase::new("b", "B", vec!["echo b".into()]));
-        assert_eq!(tester.test_count(), 2);
-    }
-
-    #[test]
-    fn run_test_unknown_id_errors() {
-        let dir = tempfile::tempdir().expect("tempdir");
-        let mut tester = SimpleTUITester::new(dir.path()).expect("new");
-        let res = tester.run_test("nope");
-        assert!(res.is_err());
-    }
-
-    #[test]
-    fn run_test_echo_succeeds() {
-        let dir = tempfile::tempdir().expect("tempdir");
-        let mut tester = SimpleTUITester::new(dir.path()).expect("new");
-        tester.set_force_subprocess(true);
-        tester.add_test(TUITestCase::new("echo", "echo test", vec!["echo hello".into()]));
-
-        let res = tester.run_test("echo").expect("run");
-        assert!(res.is_passed(), "message: {}", res.message);
-    }
-
-    #[test]
-    fn run_test_bad_command_fails() {
-        let dir = tempfile::tempdir().expect("tempdir");
-        let mut tester = SimpleTUITester::new(dir.path()).expect("new");
-        tester.set_force_subprocess(true);
-        tester.add_test(TUITestCase::new(
-            "bad",
-            "bad cmd",
-            vec!["this_command_does_not_exist_xyz".into()],
-        ));
-
-        let res = tester.run_test("bad").expect("run");
-        assert!(!res.is_passed());
-        assert!(res.message.contains("not found"), "msg: {}", res.message);
-    }
-
-    #[test]
-    fn run_test_empty_command_fails() {
-        let dir = tempfile::tempdir().expect("tempdir");
-        let mut tester = SimpleTUITester::new(dir.path()).expect("new");
-        tester.set_force_subprocess(true);
-        tester.add_test(TUITestCase::new("empty", "empty", vec!["".into()]));
-
-        let res = tester.run_test("empty").expect("run");
-        assert!(!res.is_passed());
-        assert!(res.message.contains("Empty command"), "msg: {}", res.message);
-    }
-
-    #[test]
-    fn run_all_collects_results() {
-        let dir = tempfile::tempdir().expect("tempdir");
-        let mut tester = SimpleTUITester::new(dir.path()).expect("new");
-        tester.set_force_subprocess(true);
-        tester.add_test(TUITestCase::new("a", "A", vec!["echo a".into()]));
-        tester.add_test(TUITestCase::new("b", "B", vec!["echo b".into()]));
-
-        let results = tester.run_all();
-        assert_eq!(results.len(), 2);
-        assert!(results["a"].is_passed());
-        assert!(results["b"].is_passed());
-    }
-
-    #[test]
-    fn create_amplihack_test_helper() {
-        let tc = create_amplihack_test("help", "--help");
-        assert_eq!(tc.test_id, "help");
-        assert!(tc.commands[0].contains("amplihack --help"));
-    }
-
-    #[test]
-    fn output_dir_is_created() {
-        let dir = tempfile::tempdir().expect("tempdir");
-        let nested = dir.path().join("nested").join("deep");
-        let tester = SimpleTUITester::new(&nested).expect("new");
-        assert!(tester.output_dir().exists());
-    }
-
-    #[test]
-    fn command_exists_on_path_echo() {
-        assert!(command_exists_on_path("echo"));
-    }
-
-    #[test]
-    fn command_exists_on_path_missing() {
-        assert!(!command_exists_on_path("no_such_binary_abc_xyz_123"));
-    }
-}
+#[path = "tests/simple_tui_tests.rs"]
+mod tests;
